forecast with animation and current weather for Kyiv added

+++ Lab2/WeatherInformer/WeatherClient/Models/Forecast5days3hours.cs
@@ -55,7 +55,7 @@ namespace WeatherInformer.WeatherClient.Models.Forecast5days3hours
 
     public class List
     {
-        public int dt { get; set; }
+        public long dt { get; set; }
         public Main main { get; set; }
         public List<Weather> weather { get; set; }
         public Clouds clouds { get; set; }
